requirement.txt
functions-framework==3.*
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
flask
google-genai>=0.7.0
google-cloud-storage
pyjwt[crypto]